+++ src/main/java/hillelee/pet/ErrorBody.java
@@ -1,0 +1,11 @@
+package hillelee.pet;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+class ErrorBody {
+    private String message;
+    private final Integer code = 400;
+}
