+++ src/main/java/hillelee/pet/NoSuchPetException.java
@@ -1,0 +1,9 @@
+package hillelee.pet;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+class NoSuchPetException extends RuntimeException {
+
+}
