+++ src/main/java/hillelee/pet/Pet.java
@@ -1,0 +1,14 @@
+package hillelee.pet;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+class Pet {
+    private String name;
+    private String specie;
+    private Integer age;
+}
